@@ -1,104 +1,90 @@
-import React from 'react'
-import update from 'immutability-helper'
-import Videocard from '../videocard/videocard'
+import React, { useState, useRef, useEffect } from 'react'
+import Videocard from '../videocard'
 import { DropTarget } from 'react-dnd'
 
-class Playlist extends React.Component {
-  constructor (props) {
-    super(props)
-    this.state = {
-      videos: []
-    }
-  }
+const Playlist = props => {
+  const [videos, setVideos] = useState([])
 
-  UNSAFE_componentWillUpdate (nextProps) {
+  const isFirstRender = useRef(true)
+  useEffect(() => {
+    if (isFirstRender.current) {
+      isFirstRender.current = false
+      return
+    }
     if (
-      nextProps.videoDetailsList &&
-      nextProps.videoDetailsList.length !== this.state.videos.length
+      props.videoDetailsList &&
+      props.videoDetailsList.length !== videos.length
     ) {
-      this.setState({ videos: nextProps.videoDetailsList })
+      setVideos(props.videoDetailsList)
     }
-  }
+  })
 
-  pushCard (card) {
-    this.setState(
-      update(this.state, {
-        videos: {
-          $push: [card]
-        }
-      })
+  const moveCard = (dragIndex, hoverIndex) => {
+    const newVideos = videos
+    const dragCard = newVideos[dragIndex]
+
+    newVideos.splice(dragIndex, 1)
+    newVideos.splice(hoverIndex, 0, dragCard)
+
+    setVideos(
+      newVideos
     )
+
+    props.updateVideoListOrder(videos)
   }
 
-  moveCard (dragIndex, hoverIndex) {
-    const { videos } = this.state
-    const dragCard = videos[dragIndex]
-
-    this.setState(
-      update(this.state, {
-        videos: {
-          $splice: [[dragIndex, 1], [hoverIndex, 0, dragCard]]
-        }
-      })
-    )
-    this.props.updateVideoListOrder(this.state.videos)
-  }
-
-  handleDragOver (event) {
+  const handleDragOver = (event) => {
     event.stopPropagation()
   }
 
-  render () {
-    const { videos } = this.state
-
-    const FetchSpinner = () => {
-      if (this.props.fetchInProgress) {
-        return (
-          <div className='fetchSpinnerDiv'>
-            <div class='cube-container'>
-              <div id='cube'>
-                <div class='front' />
-                <div class='back' />
-                <div class='right' />
-                <div class='left' />
-                <div class='top' />
-                <div class='bottom' />
-              </div>
-              <div id='shadow'>,</div>
+  const FetchSpinner = () => {
+    if (props.fetchInProgress) {
+      return (
+        <div className='fetchSpinnerDiv'>
+          <div class='cube-container'>
+            <div id='cube'>
+              <div class='front' />
+              <div class='back' />
+              <div class='right' />
+              <div class='left' />
+              <div class='top' />
+              <div class='bottom' />
             </div>
-
+            <div id='shadow'>,</div>
           </div>
-        )
-      } else {
-        return null
-      }
-    }
 
-    return (
-      <span onDragOver={this.handleDragOver} className='playlist-container'>
-        {videos &&
-          videos.map((video, index) => (
-            <Videocard
-              key={video.id}
-              id={video.id}
-              index={index}
-              title={video.title}
-              card={video}
-              listId={1}
-              channel={video.channel}
-              thumbnail={video.thumb}
-              onRemove={() => this.props.onRemove(video.id)}
-              moveCard={this.moveCard.bind(this)}
-            />
-          ))}
-        <FetchSpinner />
-      </span>
-    )
+        </div>
+      )
+    } else {
+      return null
+    }
   }
+
+  console.log(videos)
+  return (
+    <span onDragOver={handleDragOver} className='playlist-container'>
+      {videos &&
+        videos.map((video, index) => (
+          <Videocard
+            key={video.id}
+            id={video.id}
+            index={index}
+            title={video.title}
+            card={video}
+            listId={1}
+            channel={video.channel}
+            thumbnail={video.thumb}
+            onRemove={() => props.onRemove(video.id)}
+            moveCard={moveCard}
+          />
+        ))}
+      <FetchSpinner />
+    </span>
+  )
 }
 
 const cardTarget = {
-  drop (props) {
+  drop(props) {
     return {
       listId: props.id
     }
